requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
idna==2.8
multidict==4.7.4
pymongo==3.9.0
python-dotenv==0.10.5
slackclient==2.5.0
yarl==1.4.2
